Refactor map.rs
